🚚 move the ufunc-related callable protocols to `numpy._core.umath`
