+++ src/numpy-stubs/core/_multiarray_umath.pyi
@@ -35,84 +35,41 @@ from numpy._core._multiarray_umath import (
     __cpu_targets_info__ as __cpu_targets_info__,
     __version__ as __version__,
     _add_newdoc_ufunc as _add_newdoc_ufunc,
-    _arg as _arg,
-    _center as _center,
-    _expandtabs as _expandtabs,
-    _expandtabs_length as _expandtabs_length,
     _extobj_contextvar as _extobj_contextvar,
     _flagdict as _flagdict,
     _get_extobj_dict as _get_extobj_dict,
     _get_madvise_hugepage as _get_madvise_hugepage,
     _get_ndarray_c_version as _get_ndarray_c_version,
-    _ljust as _ljust,
-    _lstrip_chars as _lstrip_chars,
-    _lstrip_whitespace as _lstrip_whitespace,
     _make_extobj as _make_extobj,
     _monotonicity as _monotonicity,
-    _ones_like as _ones_like,
-    _partition as _partition,
-    _partition_index as _partition_index,
     _place as _place,
     _reconstruct as _reconstruct,
-    _replace as _replace,
-    _rjust as _rjust,
-    _rpartition as _rpartition,
-    _rpartition_index as _rpartition_index,
-    _rstrip_chars as _rstrip_chars,
-    _rstrip_whitespace as _rstrip_whitespace,
     _set_madvise_hugepage as _set_madvise_hugepage,
-    _strip_chars as _strip_chars,
-    _strip_whitespace as _strip_whitespace,
     _vec_string as _vec_string,
-    _zfill as _zfill,
-    absolute as absolute,
-    add as add,
     add_docstring as add_docstring,
     arange as arange,
-    arccos as arccos,
-    arccosh as arccosh,
-    arcsin as arcsin,
-    arcsinh as arcsinh,
-    arctan as arctan,
-    arctan2 as arctan2,
-    arctanh as arctanh,
     array as array,
     asanyarray as asanyarray,
     asarray as asarray,
     ascontiguousarray as ascontiguousarray,
     asfortranarray as asfortranarray,
     bincount as bincount,
-    bitwise_and as bitwise_and,
-    bitwise_count as bitwise_count,
-    bitwise_or as bitwise_or,
-    bitwise_xor as bitwise_xor,
     broadcast as broadcast,
     busday_count as busday_count,
     busday_offset as busday_offset,
     busdaycalendar as busdaycalendar,
     c_einsum as c_einsum,
     can_cast as can_cast,
-    cbrt as cbrt,
-    ceil as ceil,
     clip as clip,
     compare_chararrays as compare_chararrays,
     concatenate as concatenate,
-    conj as conj,
-    conjugate as conjugate,
-    copysign as copysign,
     copyto as copyto,
     correlate as correlate,
     correlate2 as correlate2,
-    cos as cos,
-    cosh as cosh,
     count as count,
     count_nonzero as count_nonzero,
     datetime_as_string as datetime_as_string,
     datetime_data as datetime_data,
-    deg2rad as deg2rad,
-    degrees as degrees,
-    divide as divide,
-    divmod as divmod,
     dot as dot,
     dragon4_positional as dragon4_positional,
     dragon4_scientific as dragon4_scientific,
@@ -120,126 +77,47 @@ from numpy._core._multiarray_umath import (
     empty as empty,
     empty_like as empty_like,
     endswith as endswith,
-    equal as equal,
     error as error,
     euler_gamma as euler_gamma,
-    exp as exp,
-    exp2 as exp2,
-    expm1 as expm1,
-    fabs as fabs,
     find as find,
     flagsobj as flagsobj,
     flatiter as flatiter,
-    float_power as float_power,
-    floor as floor,
-    floor_divide as floor_divide,
-    fmax as fmax,
-    fmin as fmin,
-    fmod as fmod,
     format_longfloat as format_longfloat,
-    frexp as frexp,
     from_dlpack as from_dlpack,
     frombuffer as frombuffer,
     fromfile as fromfile,
     fromiter as fromiter,
-    frompyfunc as frompyfunc,
     fromstring as fromstring,
-    gcd as gcd,
     get_handler_name as get_handler_name,
     get_handler_version as get_handler_version,
-    greater as greater,
-    greater_equal as greater_equal,
-    heaviside as heaviside,
-    hypot as hypot,
     index as index,
     inner as inner,
     interp as interp,
     interp_complex as interp_complex,
-    invert as invert,
     is_busday as is_busday,
-    isalnum as isalnum,
-    isalpha as isalpha,
-    isdecimal as isdecimal,
-    isdigit as isdigit,
-    isfinite as isfinite,
-    isinf as isinf,
-    islower as islower,
-    isnan as isnan,
-    isnat as isnat,
-    isnumeric as isnumeric,
-    isspace as isspace,
-    istitle as istitle,
-    isupper as isupper,
-    lcm as lcm,
-    ldexp as ldexp,
-    left_shift as left_shift,
-    less as less,
-    less_equal as less_equal,
     lexsort as lexsort,
-    log as log,
-    log1p as log1p,
-    log2 as log2,
-    log10 as log10,
-    logaddexp as logaddexp,
-    logaddexp2 as logaddexp2,
-    logical_and as logical_and,
-    logical_not as logical_not,
-    logical_or as logical_or,
-    logical_xor as logical_xor,
-    matvec as matvec,
-    maximum as maximum,
     may_share_memory as may_share_memory,
     min_scalar_type as min_scalar_type,
-    minimum as minimum,
-    mod as mod,
-    modf as modf,
-    multiply as multiply,
     nditer as nditer,
-    negative as negative,
     nested_iters as nested_iters,
-    nextafter as nextafter,
-    not_equal as not_equal,
     packbits as packbits,
     pi as pi,
-    positive as positive,
-    power as power,
     promote_types as promote_types,
     putmask as putmask,
-    rad2deg as rad2deg,
-    radians as radians,
     ravel_multi_index as ravel_multi_index,
-    reciprocal as reciprocal,
-    remainder as remainder,
     result_type as result_type,
     rfind as rfind,
-    right_shift as right_shift,
     rindex as rindex,
-    rint as rint,
     scalar as scalar,
     set_datetimeparse_function as set_datetimeparse_function,
     set_typeDict as set_typeDict,
     shares_memory as shares_memory,
-    sign as sign,
-    signbit as signbit,
-    sin as sin,
-    sinh as sinh,
-    spacing as spacing,
-    sqrt as sqrt,
-    square as square,
     startswith as startswith,
-    str_len as str_len,
-    subtract as subtract,
-    tan as tan,
-    tanh as tanh,
     tracemalloc_domain as tracemalloc_domain,
-    true_divide as true_divide,
-    trunc as trunc,
     typeinfo as typeinfo,
     unpackbits as unpackbits,
     unravel_index as unravel_index,
     vdot as vdot,
-    vecdot as vecdot,
-    vecmat as vecmat,
     where as where,
     zeros as zeros,
 )
